pyproject.toml
[tool.poetry]
name = "pydantic-gen"
version = "0.3.0"
description = "Code generator for pydantic schemas"
authors = ["MB <mb@blaster.ai>"]
license = "MIT"
readme='README.md'
homepage='https://github.com/licht1stein/pydantic-gen'

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.4"
jinja2 = "^2.11.2"
attrs = "^19.3.0"
python-box = "^4.2.2"
"ruamel.yaml" = "^0.16.10"
black = "^19.10b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
coverage = "^5.1"
pytest-cov = "^2.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
